test: the fence-rendering assertions move to the markdown suite, which lives in MeshWeaver.Plugins (#2511)
`MeshWeaver.Markdown` is compiled HERE and tested THERE. This repo has no test project referencing
`MeshWeaver.Markdown` at all; `MeshWeaver.Markdown.Test` lives in MeshWeaver.Plugins, next to
`ExecutableCodeCellRenderingTest`, `MermaidRenderingTest` and `MarkdownFenceEditingTest`, and
project-references this repo through `$(MeshWeaverRoot)`. `PromptFenceComposerTest` was written in
`MeshWeaver.Graph.Test` because that is where the layout area lives — the wrong home for a renderer
test, and one that makes this repo's green look like it covers a fence's HTML when nothing here
does.

The nine rendering assertions land in Systemorph/MeshWeaver.Plugins#1181 unchanged. What stays is
`PromptComposerAreaTest`: the `Prompt` area's own shape — the compact `ThreadChatControl` carrying
the decoded draft — which is this repo's code and whose entry point (`MeshNodeLayoutAreas.
PromptComposer`) is internal here. Keeping it avoids a new cross-repo `InternalsVisibleTo`, a drift
hazard this very csproj already documents at length.

`Doc/Architecture/MarkdownFenceExtensions` now names both homes and says why, so the next fence
author does not write a rendering assertion into a repo that cannot run one.

Verified: `dotnet build -c Release -warnaserror` 0 Error(s) / 0 Warning(s); full unfiltered
`MeshWeaver.Graph.Test` — 1207 passed, 0 failed.
